Refactor UnitTests namespaces to include"Catalog." prefix in the name
