+++ Catalog/Catalog.UnitTests/Catalog.Api/Mappers/BrandMapperTests.cs
@@ -4,7 +4,7 @@ using Catalog.Domain.Entities;
 using FluentAssertions;
 using Xunit;
 
-namespace Catalog.UnitTests.Api.Mappers;
+namespace Catalog.UnitTests.Catalog.Api.Mappers;
 
 public class BrandMapperTests
 {
