+++ Catalog/Catalog.UnitTests/Catalog.Api/Mappers/CategoryMapperTests.cs
@@ -4,7 +4,7 @@ using Catalog.Domain.Entities;
 using FluentAssertions;
 using Xunit;
 
-namespace Catalog.UnitTests.Api.Mappers;
+namespace Catalog.UnitTests.Catalog.Api.Mappers;
 
 public class CategoryMapperTests
 {
