+++ Catalog/Catalog.UnitTests/Catalog.UnitTests.csproj
@@ -27,4 +27,8 @@
       <ProjectReference Include="..\Catalog.IntegrationTests\Catalog.IntegrationTests.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="TestUtils\" />
+    </ItemGroup>
+
 </Project>
